revised tests

+++ app/assets/javascripts/controllers/lesson/lesson_index_controller.js
@@ -1,8 +1,23 @@
 App.LessonIndexController = Ember.ObjectController.extend({
+
+	needs: ['course'],
+
 	actions: {
 		delete: function(lesson) {
-			lesson.destroyRecord();
-			this.transitionToRoute('lessons.index');
+			var userConfirm = confirm("Are you sure you want to delete the lesson '" + lesson.get('name') + "'?");
+			if (userConfirm) {
+				lesson.destroyRecord();
+				this.transitionToRoute('lessons');
+			}
+			else {
+				this.transitionToRoute('lesson', lesson);
+			}
+		},
+
+		addQuestion: function(lesson) {
+			var question = this.store.createRecord('question', {
+				
+			})
 		}
 	}
 });
